new method `to_md` for class `Markdown`, new function `to_md_parser` for repo; more tests; better repo structure
This commit come with a better structure for the repository. It provides a new method `to_md` for the `Markdown` class which makes it possible to parse the data back to markdown formatted string. The method comes with the possibilities to include or exclude certain markdown elements. It also come with the possiblities to set a distinct spacer (= empty paragraph) between each parsed markdown elemen.